Expand group ids to 64 bits
Changed group ids from uint_32 to uint_64 to allow higher group numbers
and avoid overflow when solving large composite groups.

+++ src/sketch.h
@@ -53,7 +53,7 @@ enum class Command : uint32_t;
 class hGroup {
 public:
     // bits 15: 0   -- group index
-    uint32_t v;
+    uint64_t v;
 
     inline hEntity entity(int i) const;
     inline hParam param(int i) const;
